% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhawkes.R
\name{mhawkes}
\alias{mhawkes}
\title{Simulation of a multivariate Hawkes process}
\usage{
mhawkes(tMax, eta, repr, beta, keep_ancestors = TRUE)
}
\arguments{
\item{tMax}{A double; the end time of the observation window \link{0,T}}

\item{eta}{A vector of non-negative doubles; the immigration parameters}

\item{repr}{A matrix of non-negative doubles; the mean numbers of offsprings}

\item{beta}{A matrix of non-negative doubles; the rates of the exponential reproduction functions}

\item{keep_ancestors}{A boolean; track ancestors (used only when plotting the process)}
}
\value{
A S3 object of class \code{mhawkes} containing a vector (\code{.$pointproc}) of simulated values, and all other objects used for the simulation.
}
\description{
Simulates a multivariate Hawkes process with exponential reproduction functions. The argument \code{keep_ancestors} is useful when plotting the patterns with \code{plot_mhawkes}, but slows the computation. It should be set to \code{FALSE} if only the simulation matters (and not the cluster structure).
}
\examples{
tMax = 10
eta = c(2, 1, 3)
repr = matrix(runif(9, 0, .33), 3, 3)
beta = matrix(rexp(9, 1), 3, 3)
x = mhawkes(tMax, eta, repr, beta)
plot(x)
}
