% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley.R
\name{mripley}
\alias{mripley}
\title{K-Ripley function for multivariate point patterns}
\usage{
mripley(pp, tMax, rmin = 0.1, rmax = 2, step = 0.1)
}
\arguments{
\item{pp}{A list of vectors of ordered time points; the point patterns}

\item{tMax}{A double; the end time of the observation window \link{0,T}}

\item{rmin}{A double; the minimal radius at which to compute the K-function}

\item{rmax}{A double; the maximal radius at which to compute the K-function}

\item{step}{A double; the successive increases between rmin and rmax}
}
\value{
A tibble object with columns \code{i}, \code{j}, \code{r}, and \code{ripley}; the value \code{ripley} of the K-function between components \code{i} and \code{j} at each corresponding radius \code{r}
}
\description{
Computes the K-Ripley function for a multivariate point pattern
}
\examples{
tMax = 1000
eta = c(2, 1, 3)
repr = matrix(runif(9, 0, 1), 3, 3) \%>\% {runif(1, .3, .7) * . / invisible(eigen(.))$values[1]} \%>\% Re()
beta = matrix(rexp(9, 1), 3, 3)
x = mhawkes(tMax, eta, repr, beta, keep_ancestors = FALSE)
x_ripley = mripley(x$pointproc, tMax)
ggplot2::ggplot(x_ripley) +
  ggplot2::aes(x = r, y = ripley - r) +
  ggplot2::geom_line() +
  ggplot2::facet_grid(i ~ j)
}
