% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhawkes.R
\name{plot.mhawkes}
\alias{plot.mhawkes}
\title{Plot of a multivariate Hawkes process}
\usage{
\method{plot}{mhawkes}(x)
}
\arguments{
\item{x}{An object of class \code{mhawkes}, generated by function \code{mhawkes}.}
}
\value{
None
}
\description{
Plots the realisation of a multivariate Hawkes process according to its cluster representation. Only works with objects generated from \code{mhawkes} with argument \code{keep_ancestors} set to \code{TRUE}.
}
