% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley-package.R
\docType{package}
\name{ripley-package}
\alias{ripley-package}
\title{ripley: Efficient K-Ripley Function}
\description{
Computes the K-Ripley function of a point pattern in almost linear time.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
