% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripley.R
\name{ripley}
\alias{ripley}
\title{K-Ripley function}
\usage{
ripley(pp, tMax, rmin = 0.1, rmax = 2, step = 0.1)
}
\arguments{
\item{pp}{A vector of ordered time points; the point pattern}

\item{tMax}{A double; the end time of the observation window \link{0,T}}

\item{rmin}{A double; the minimal radius at which to compute the K-function}

\item{rmax}{A double; the maximal radius at which to compute the K-function}

\item{step}{A double; the successive increases between rmin and rmax}
}
\value{
A tibble object with columns \code{r} and \code{ripley}; the value of the K-function at each corresponding radius
}
\description{
Computes the K-Ripley function for a point pattern
}
\examples{
tMax = 1000
x = hawkesbow::hawkes(tMax, fun=1, repr=0.5, family="exp", rate=2)
x_ripley = ripley(x$p, tMax)
ggplot2::ggplot(x_ripley) +
  ggplot2::aes(x = r, y = ripley - r) +
  ggplot2::geom_line()
}
